+++ src/FluentDbTools/Abstractions/FluentDbTools.Migration.Abstractions/IMigratorResolver.cs
@@ -1,7 +1,0 @@
-﻿namespace FluentDbTools.Migration.Abstractions
-{
-    public interface IMigrationResolver : IMigrationExecutor
-    {
-        bool CanMigrate();
-    }
-}