+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- the topmost header version must be set manually in the VERSION file -->
+### Version 1.0.3 2019-04-01
+ - added SchemaPassword property to IDbConfig
+ - Deleted old stuff from FluentDbTools.Migration.Abstractions 
+ - Removed FluentDbTools.Database.Abstractions assembly
+
 ### Version 1.0.2 2019-01-31
  - Updated nuget metadata.